refactor: define cli exports
